requirements.txt
langgraph>=0.3
langchain>=0.3
langchain-openai>=0.3
aiosqlite>=0.20
sqlalchemy>=2.0
pydantic>=2.0
python-dotenv>=1.0
pyyaml>=6.0
langgraph-checkpoint-sqlite>=3.0
pytest>=8.0
pytest-asyncio>=0.25
